refactor: 親インターフェイスを抽象クラスに変更。汎用メソッドを親に移動。
