+++ src/components/diagrams/icon/GenericSingleIconGenerator.ts
@@ -5,8 +5,9 @@ import IconFontAndChar from '@/components/diagrams/icon/IconFontAndChar';
 import Resource from '@/domain/resource/Resource';
 import ResourceType from '@/domain/resource/ResourceType';
 import Placement from "@/domain/diagram/placement/Placement";
+import { extend } from 'jquery';
 
-export default abstract class GenericSingleIconGenerator implements IconGenerator<Resource> {
+export default abstract class GenericSingleIconGenerator extends IconGenerator<Resource> {
     public abstract resourceType(): ResourceType;
 
     public generate(placement: Placement, resource: Resource, iconChar: IconFontAndChar): Figure {
@@ -35,10 +36,7 @@ export default abstract class GenericSingleIconGenerator implements IconGenerato
             selectable: false
         });
 
-        icon.createPort(
-            "hybrid",
-            new draw2d.layout.locator.CenterLocator()
-        );
+        icon.createPort("hybrid", new draw2d.layout.locator.CenterLocator());
         // PortからではなくFigureから線が出ているように見せるため、アンカー設定。
         const port = icon.getPorts().last();
         const anchor = new draw2d.layout.anchor.ChopboxConnectionAnchor(icon);
